tests: change BigInteger to BigDecimal and increase totalSizeByteLimit to 375 for CoverageRetentionTest

+++ test2code/src/test/kotlin/com/epam/drill/test2code/coverage/InMemoryRetentionQueueTest.kt
@@ -1,3 +1,18 @@
+/**
+ * Copyright 2020 - 2022 EPAM Systems
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.epam.drill.test2code.coverage
 
 import org.junit.jupiter.api.*
@@ -10,7 +25,7 @@ class InMemoryRetentionQueueTest {
         val byteCount = 6
         val innerArraySize = 2
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
-        val sizeLimit = BigInteger.valueOf(10)
+        val sizeLimit = BigDecimal.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
 
         queue.addAll(bytes)
@@ -24,7 +39,7 @@ class InMemoryRetentionQueueTest {
         val byteCount = 1
         val innerArraySize = 10
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
-        val sizeLimit = BigInteger.valueOf(10)
+        val sizeLimit = BigDecimal.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
 
         queue.addAll(bytes)
@@ -35,7 +50,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `adding single element larger than limit should result in an empty queue`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
         val byteCount = 1
         val innerArraySize = 12
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
@@ -47,7 +62,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `flush should return the same elements as added to the queue`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
         val byteCount = 1
         val innerArraySize = 5
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
@@ -60,7 +75,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `queue after flush must be empty`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
         val byteCount = 1
         val innerArraySize = 5
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
@@ -75,7 +90,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `queue should only fit elements with size less than or equal to total limit`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
         val expectedSizeOfQueue = 4
         val bytes = sequenceOf(
             ByteArray(12) { 1 },
@@ -91,7 +106,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `queue should contain elements within limit and exclude element exceeding limit `() {
-        val sizeLimit = BigInteger.valueOf(10)
+        val sizeLimit = BigDecimal.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
         val expectedBytes1 = ByteArray(2) { 1 }
         val expectedBytes2 = ByteArray(2) { 2 }
@@ -116,7 +131,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `adding new elements after flush should be reflected in queue content`() {
-        val sizeLimit = BigInteger.valueOf(10)
+        val sizeLimit = BigDecimal.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
         val expectedBytes1 = ByteArray(2) { 1 }
         val expectedBytes2 = ByteArray(2) { 2 }
@@ -139,7 +154,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `initializing queue with negative totalSizeByteLimit should throw exception`() {
-        val sizeLimit = BigInteger.valueOf(-10)
+        val sizeLimit = BigDecimal.valueOf(-10)
 
         assertThrows<IllegalArgumentException> {
             val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
@@ -149,7 +164,7 @@ class InMemoryRetentionQueueTest {
     @Test
     fun `adding elements to a queue with zero totalSizeByteLimit should return an empty queue`() {
         val expectedSizeOfQueue = 0
-        val sizeLimit = BigInteger.ZERO
+        val sizeLimit = BigDecimal.ZERO
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
         val bytes = sequenceOf(ByteArray(1) { 1 })
